Simplify serdes config by merging serde config and schema resolver config
