+++ docs/modules/ROOT/partials/getting-started/proc-registry-serdes-config-consumer.adoc
@@ -20,7 +20,7 @@ This procedure describes how to configure a Kafka consumer client written in Jav
 ----
 String registryUrl = "https://registry.example.com/apis/registry/v2";
 Properties props = new Properties();
-props.putIfAbsent(SerdeConfig.REGISTRY_URL, registryUrl); 
+props.putIfAbsent(SerdeConfig.REGISTRY_URL, registryUrl);
 ----
 
 . Configure the client with the {registry} deserializer. For example:
