+++ docs/modules/ROOT/partials/getting-started/ref-registry-serdes-config-props.adoc
@@ -58,19 +58,9 @@ The `DefaultSchemaResolver` provides the following properties to configure acces
 |Used by serializers and deserializers. URL to access the registry API.
 |`String`
 |None
-|`AUTH_SERVICE_URL`
-|`apicurio.auth.service.url`
-|Used by serializers and deserializers. URL of the authentication service. Required when accessing a secure registry using the OAuth client credentials flow.
-|`String`
-|None
 |`AUTH_TOKEN_ENDPOINT`
 |`apicurio.auth.service.token.endpoint`
-|Used by serializers and deserializers. URL of the token endpoint. Required when accessing a secure registry and `AUTH_SERVICE_URL` is not specified.
-|`String`
-|None
-|`AUTH_REALM`
-|`apicurio.auth.realm`
-|Used by serializers and deserializers. Realm to access the authentication service. Required when accessing a secure registry using the OAuth client credentials flow.
+|Used by serializers and deserializers. URL of the token endpoint.
 |`String`
 |None
 |`AUTH_CLIENT_ID`
